Updated to .Net10 and dalamud 14, changed adapters to extensions due to dalamud14 incompadtabilities

+++ Collections/Collectibles/Collectible/EmoteCollectible.cs
@@ -46,7 +46,7 @@ public class EmoteCollectible : Collectible<Emote>, ICreateable<EmoteCollectible
 
     protected override int GetIconId()
     {
-        return ExcelRow.Icon;
+        return (int)ExcelRow.Icon;
     }
 
     public override unsafe void Interact()
